Made the Spell Cast affix use a min and max field instead of StepFunction

+++ src/main/java/net/kayn/fallen_gems_affixes/event/SpellEventHandler.java
@@ -48,7 +48,7 @@ public class SpellEventHandler {
                 } else if (inst.affix().get() instanceof SpellCastAffix affix &&
                         affix.trigger == SpellCastAffix.TriggerType.SPELL_DAMAGE) {
                     LivingEntity actualTarget = getSpellTarget(affix.target, caster, target);
-                    affix.triggerSpell(caster, actualTarget, inst.rarity().get(), inst.level());
+                    affix.triggerSpell(caster, actualTarget, inst.rarity().get(), (int) inst.level());
                 }
             });
 
@@ -82,7 +82,7 @@ public class SpellEventHandler {
                 } else if (inst.affix().get() instanceof SpellCastAffix affix &&
                         affix.trigger == SpellCastAffix.TriggerType.SPELL_HEAL) {
                     LivingEntity actualTarget = getSpellTarget(affix.target, caster, healTarget);
-                    affix.triggerSpell(caster, actualTarget, inst.rarity().get(), inst.level());
+                    affix.triggerSpell(caster, actualTarget, inst.rarity().get(), (int) inst.level());
                 }
             });
